+++ src/edu/antonym/prototype/VectorEmbedding.java
@@ -3,7 +3,7 @@ package edu.antonym.prototype;
 public interface VectorEmbedding {
 
 	int getDimension();
-	public float[] getVectorRep(int word);
-	public int findClosestWord(float[] vector);
+	public double[] getVectorRep(int word);
+	public int findClosestWord(double[] vector);
 	Vocabulary getVocab();
 }
